kbfs: enable stricter golangci linting
And pin the linter to a particular version.

Suggested by joshblum.

Issue: #18565

+++ go/kbfs/data/bsplitter_simple.go
@@ -256,8 +256,8 @@ func (b *BlockSplitterSimple) SplitDirIfNeeded(block *DirBlock) (
 	// Delete the second half of the names from the original block,
 	// and add to the new block.
 	newBlock := NewDirBlock().(*DirBlock)
-	startOff := int(len(names) / 2)
-	for _, name := range names[int(len(names)/2):] {
+	startOff := len(names) / 2
+	for _, name := range names[len(names)/2:] {
 		newBlock.Children[name] = block.Children[name]
 		delete(block.Children, name)
 	}
